Refactor header files and update includes.
Renamed all `.h` files to `.hpp` to adhere to C++ conventions. Updated include directives in the corresponding source files to match the new header filenames.
